+++ contracts/src/token/test/simulators/NativeShieldedTokenSimulator.ts
@@ -3,8 +3,6 @@ import {
   type SimulatorOptions,
 } from '@openzeppelin/compact-simulator';
 import {
-  type ContractAddress,
-  type Either,
   ledger,
   type Maybe,
   Contract as MockNativeShieldedToken,
@@ -119,11 +117,23 @@ export class NativeShieldedTokenSimulator extends NativeShieldedTokenSimulatorBa
    * @returns The newly created coin's info (nonce, color, value).
    */
   public _mint(
-    recipient: Either<ZswapCoinPublicKey, ContractAddress>,
+    recipient: ZswapCoinPublicKey,
     amount: bigint,
     nonce: Uint8Array,
   ): Promise<ShieldedCoinInfo> {
     return this.circuits.impure._mint(recipient, amount, nonce);
+  }
+
+  /**
+   * @description Mints `amount` to the contract itself, claimed in the same
+   * call.
+   * @returns The newly created coin's info (nonce, color, value).
+   */
+  public _mintToSelf(
+    amount: bigint,
+    nonce: Uint8Array,
+  ): Promise<ShieldedCoinInfo> {
+    return this.circuits.impure._mintToSelf(amount, nonce);
   }
 
   /**
@@ -134,7 +144,7 @@ export class NativeShieldedTokenSimulator extends NativeShieldedTokenSimulatorBa
   public _burn(
     coin: ShieldedCoinInfo,
     amount: bigint,
-    refundTo: Either<ZswapCoinPublicKey, ContractAddress>,
+    refundTo: ZswapCoinPublicKey,
   ): Promise<Maybe<ShieldedCoinInfo>> {
     return this.circuits.impure._burn(coin, amount, refundTo);
   }